Split urlLabels into own object, add index.js to labes to help with imports
