+++ src/Labels/index.js
@@ -1,16 +1,2 @@
-const labels = {
-  findRestaurants: "Find restaurants near you:",
-  header: "Rstrnt Findr-r",
-  findByCuisine: "Find By Cuisine",
-  findButton: "Find",
-  searchResultQuery: "Restaurants near ",
-  searchResultURL: "/searchresults",
-  loading: "Loading...",
-  orderBy: "Order by:",
-  nearestFilter: "Nearest to you",
-  ratingFilter: "By rating",
-  seeRestaurantButton: "See restaurant",
-  restaurantDetailURL: "/restaurant/",
-};
-
-export { labels };
+export * from "./textLabels";
+export * from "./urlLabels";
