+++ src/Labels/textLabels.js
@@ -1,0 +1,14 @@
+const labels = {
+  findRestaurants: "Find restaurants near you:",
+  header: "Rstrnt Findr-r",
+  findByCuisine: "Find By Cuisine",
+  findButton: "Find",
+  searchResultQuery: "Restaurants near ",
+  loading: "Loading...",
+  orderBy: "Order by:",
+  nearestFilter: "Nearest to you",
+  ratingFilter: "By rating",
+  seeRestaurantButton: "See restaurant",
+};
+
+export { labels };
