+++ src/Labels/urlLabels.js
@@ -1,0 +1,6 @@
+const urlLabels = {
+  searchResultURL: "/searchresults",
+  restaurantDetailURL: "/restaurant/",
+};
+
+export { urlLabels };
